Refactor code to fix log out bug

+++ bucket_app/views.py
@@ -48,7 +48,8 @@ def create_bucket_list():
     """
     View that renders the bucket list  creation form
     """
-
+    if not auth.current_user.is_logged_in:
+        return redirect('/')
 
     form = CreateBucketListForm(csrf_enabled=False)
 
@@ -64,7 +65,8 @@ def create_bucket_list():
 
 @app.route('/update/<int:id>', methods=['GET', 'POST'])
 def update_bucket(id):
-
+    if not auth.current_user.is_logged_in:
+        return redirect('/')
     val = id - 1
     u_form = UpdateBucketListForm(csrf_enabled=False)
     if u_form.validate_on_submit():
@@ -101,7 +103,8 @@ def view_bucket_lists():
 
 @app.route('/update/<int:id>/<old_name>', methods=['GET', 'POST'])
 def update_bucket_item(id, old_name):
-
+    if not auth.current_user.is_logged_in:
+        return redirect('/')
 
     val = id - 1
     u_form = UpdateBucketItemForm(csrf_enabled=False)
@@ -140,7 +143,8 @@ def view_bucket_items(id):
 
 @app.route('/bucket-item/<int:id>', methods=['GET', 'POST'])
 def create_bucket_items(id):
-
+    if not auth.current_user.is_logged_in:
+        return redirect('/')
     val = id - 1
     form = CreateBucketItemForm(csrf_enabled=False)
     if form.validate_on_submit():
@@ -156,7 +160,8 @@ def delete_bucket(id):
     """
     View that renders the deletes bucket lists
     """
-    
+    if not auth.current_user.is_logged_in:
+        return redirect('/')
     val = int(id) - 1
     bucket.delete_bucket(auth.current_user.email, val)
     return redirect(request.referrer)
@@ -167,7 +172,8 @@ def delete_bucket_item(id, name):
     """
     View that  deletes bucket lists item
     """
-
+    if not auth.current_user.is_logged_in:
+        return redirect('/')
     val = int(id) - 1
     bucket.delete_bucket_item(auth.current_user.email, val, name)
     return redirect(request.referrer)
